backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pymysql==1.0.3
cryptography==40.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
alembic==1.10.4